backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
crewai==0.19.0
chromadb
psycopg2-binary
sqlalchemy
python-dotenv
pydantic>=2.0
pydantic-settings
requests
httpx
python-multipart
reportlab>=4.0.0
pypdf>=4.0.0
python-docx>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
psutil>=5.9.0
PyJWT
email-validator
